// lib/Color.d.ts
export declare enum Color {
    black = 0,
    blue = 1,
    red = 2,
    green = 3,
    yellow = 4,
    grey = 5,
    fuschia = 6,
    orange = 7,
    teal = 8,
    brown = 9,
    true_black = 10,
    not_written = 11,
    false = 0,
    true = 1,
    no_color = -1
}
export declare const identity_color_mapping: Color[];