// lib/Transform.d.ts
export declare enum Transform {
    identity = 0,
    flip_x = 1,
    flip_y = 2,
    rotate_90 = 3,
    rotate_180 = 4,
    rotate_270 = 5,
    transpose = 6,
    opposite_transpose = 7,
    tile_x = 8,
    tile_y = 9,
    tile = 10,
    ping_pong_x = 11,
    ping_pong_y = 12,
    ping_pong = 13,
    rotation = 14
}